requirements.txt
# requirements-termux.txt - Optimized for Termux ARM architecture
# Use only packages with pre-built wheels available

# Core Flask (all have wheels)
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.0.5
werkzeug==3.0.1

# Database (has wheel)
sqlalchemy==2.0.23

# Google AI (use versions with wheels)
google-generativeai==0.3.2
google-auth==2.23.4

# Image processing (has wheel)
pillow==10.1.0

# File handling (all have wheels)
python-dotenv==1.0.0
requests==2.31.0

# PDF and audio (both have wheels)
reportlab==4.0.4
gtts==2.4.0

# Dependencies with wheels
click==8.1.7
jinja2==3.1.3
markupsafe==2.1.3
itsdangerous==2.1.2
blinker==1.7.0